component declaration refactor to arrow syntax
