+++ website/src/components/Home/Home.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import logo from '../../assets/logo/logo512.png';
 import styles from './Home.module.css';
-import Game from '../Game/Game';
 
 function Home() {
   return (
     <div className={styles.root}>
-      <Game />
+      <img src={logo} className={styles.appLogo} alt="logo" />
+      <p className={styles.text}>
+        Self-Organising Multi-Agent Systems 2020
+        </p>
     </div>
   );
 }
