--- website/src/components/Visualisations.tsx
@@ -1,6 +1,0 @@
-export { default as Game } from "./Game/Game";
-export { default as Resources } from "./Resources/Resources";
-export { default as Roles } from "./Roles/Roles";
-export { default as IIGO } from "./IIGO/IIGO";
-export { default as IITO } from "./IITO/IITO";
-export { default as IIFO } from "./IIFO/IIFO";
