+++ website/src/components/visualisation/Game/Game.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 // import logo from '../../assets/logo/logo512.png';
 // import styles from './Game.module.css';
-import SampleVis from '../visualisation/SampleVis';
+// import SampleVis from '../visualisation/SampleVis';
 
-function Home() {
+const Game = () => {
   return (
     <div>
       {/* <img src={logo} className={styles.appLogo} alt="logo" /> */}
       <h2>Game Visualisation base</h2>
-      <SampleVis data={[1, 2, 3]} />
+      {/* <SampleVis data={[1, 2, 3]} /> */}
     </div>
   );
 }
 
-export default Home;
+export default Game;
