+++ website/src/components/visualisation/IIFO/IIFO.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import logo from '../../assets/logo/logo512.png';
+import logo from '../../../assets/logo/logo512.png';
 import styles from './IIFO.module.css';
 
-function IIFO() {
+const IIFO = () => {
   return (
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
