+++ website/src/components/visualisation/IIGO/IIGO.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import logo from '../../assets/logo/logo512.png';
+import logo from '../../../assets/logo/logo512.png';
 import styles from './IIGO.module.css';
 
-function IIGO() {
+const IIGO = () => {
   return (
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
