+++ website/src/components/visualisation/IITO/IITO.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import logo from '../../assets/logo/logo512.png';
+import logo from '../../../assets/logo/logo512.png';
 import styles from './IITO.module.css';
 
-function IITO() {
+const IITO = () => {
   return (
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
