+++ website/src/components/visualisation/Resources/Resources.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import logo from '../../assets/logo/logo512.png';
+import logo from '../../../assets/logo/logo512.png';
 import styles from './Resources.module.css';
 
-function Resources() {
+const Resources = () => {
   return (
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
