+++ website/src/components/visualisation/Roles/Roles.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import logo from '../../assets/logo/logo512.png';
+import logo from '../../../assets/logo/logo512.png';
 import styles from './Roles.module.css';
 
-function Roles() {
+const Roles = () => {
   return (
     <div className={styles.root}>
       <img src={logo} className={styles.appLogo} alt="logo" />
