+++ website/src/consts/paths.ts
@@ -1,6 +1,6 @@
 export const cioutput = `/cioutput`
 export const newrun = `/newrun`
-export const gamevisualisation = `/visualisation/game`
+export const gamevisualisation = `/game`
 export const iigovisualisation = `/visualisation/iigo`
 export const iitovisualisation = `/visualisation/iito`
 export const iifovisualisation = `/visualisation/iifo`
